requirements.txt
numpy
scipy
pyyaml
python-dotenv
Pillow
pdf2image